[r] Move the AesCipherParams structure to the cpp file
