+++ src/crypto_guard_ctx.cpp
@@ -3,11 +3,27 @@
 
 namespace CryptoGuard {
 
+struct AesCipherParams {
+    static const size_t KEY_SIZE = 32;             // AES-256 key size
+    static const size_t IV_SIZE = 16;              // AES block size (IV length)
+    const EVP_CIPHER *cipher = EVP_aes_256_cbc();  // Cipher algorithm
+
+    int encrypt;                              // 1 for encryption, 0 for decryption
+    std::array<unsigned char, KEY_SIZE> key;  // Encryption key
+    std::array<unsigned char, IV_SIZE> iv;    // Initialization vector
+};
+
 class CryptoGuardCtx::Impl {
 public:
+    Impl() { OpenSSL_add_all_algorithms(); }
+    ~Impl() { EVP_cleanup(); }
+
     void EncryptFile(std::iostream &inStream, std::iostream &outStream, std::string_view password);
     void DecryptFile(std::iostream &inStream, std::iostream &outStream, std::string_view password);
     std::string CalculateChecksum(std::iostream &inStream);
+
+private:
+    AesCipherParams CreateChiperParamsFromPassword(std::string_view password);
 };
 
 void CryptoGuardCtx::Impl::EncryptFile(std::iostream &inStream, std::iostream &outStream, std::string_view password) {
@@ -18,6 +34,21 @@ void CryptoGuardCtx::Impl::DecryptFile(std::iostream &inStream, std::iostream &o
     std::println("Impl DecryptFile");
 }
 std::string CryptoGuardCtx::Impl::CalculateChecksum(std::iostream &inStream) { return "Impl NOT_IMPLEMENTED"; }
+
+AesCipherParams CryptoGuardCtx::Impl::CreateChiperParamsFromPassword(std::string_view password) {
+    AesCipherParams params;
+    constexpr std::array<unsigned char, 8> salt = {'1', '2', '3', '4', '5', '6', '7', '8'};
+
+    int result = EVP_BytesToKey(params.cipher, EVP_sha256(), salt.data(),
+                                reinterpret_cast<const unsigned char *>(password.data()), password.size(), 1,
+                                params.key.data(), params.iv.data());
+
+    if (result == 0) {
+        throw std::runtime_error{"Failed to create a key from password"};
+    }
+
+    return params;
+}
 
 // --------------------------------------------------------------------------
